Move QR code from main page to header and generalize it
- Removed the hardcoded QR code from the home page (src/routes/+page.svelte) and realigned profile section layout.
- Created a static, Lucide-based QR code icon at src/components/icon/QRCode.svelte.
- Integrated a dynamic, generalized QR code generator dropdown into the desktop header (src/components/layout/Header.svelte) next to the Theme select button, with width 126px and h-6/w-6 icon size (10% smaller per PR review feedback).
- Enabled mobile-drawer QR code integration (src/components/layout/Header.svelte) displaying the QR code directly inside the open drawer list, with width 126px.
- Generalised QRCode.svelte to render canvas reactively using Svelte 5's $effect rune to update as the URL navigates.
- Handled SvelteKit page store safe-navigation fallback to window.location.href for resilient unit test execution under Happy DOM/Vitest.
- Added comprehensive Vitest unit tests in Header.test.ts, icon.test.ts and Playwright E2E tests in e2e/routes/page.test.ts utilizing testInfo.project.name for project-specific context (per PR review feedback).
- Updated Playwright E2E visual screenshots of the homepage with the new QR-free layout. All tests pass successfully.
- Fixed svelte-check and eslint warnings in Header.svelte and formatted all files with Prettier.
