+++ src/components/icon/QRCode.svelte
