+++ tests/components/icon.test.ts
@@ -4,6 +4,7 @@ import Gmail from '$components/icon/Gmail.svelte';
 import LinkedIn from '$components/icon/LinkedIn.svelte';
 import Menu from '$components/icon/Menu.svelte';
 import Moon from '$components/icon/Moon.svelte';
+import QRCode from '$components/icon/QRCode.svelte';
 import RSS from '$components/icon/RSS.svelte';
 import Search from '$components/icon/Search.svelte';
 import Sun from '$components/icon/Sun.svelte';
@@ -18,7 +19,8 @@ it.each([
 	{ Component: RSS },
 	{ Component: Moon },
 	{ Component: Search },
-	{ Component: Sun }
+	{ Component: Sun },
+	{ Component: QRCode }
 ])('render icon $Component.name', ({ Component }) => {
 	const { container } = render(Component);
 	expect(container).toBeTruthy();
